backend/requirements.txt

fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.22
asyncpg==0.27.0
psycopg2-binary==2.9.7
python-multipart==0.0.6
openai==1.3.0
langchain==0.0.302
cachetools==5.3.0
aioredis==2.0.1
redis>=4.6.0
httpx==0.24.1
pydantic==1.10.7
pandas==2.1.1
reportlab==4.0.0
python-dotenv==1.0.0
pyjwt==2.8.0
python-jose==3.3.0
cryptography==41.0.3
